+++ projects/rocprofiler-systems/tests/rocprof-sys-fork-tests.cmake
@@ -28,7 +28,7 @@
 
 rocprofiler_systems_add_test(
     NAME fork
-    TARGET fork-example
+    TARGET fork-sample
     REWRITE_ARGS -e -v 2 --print-instrumented modules -i 16
     RUNTIME_ARGS -e -v 1 --label file -i 16
     ENVIRONMENT